requirements.txt
sqlparse==0.4.4
pymongo==4.6.1
python-dotenv==1.0.0
pytest==7.4.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai>=1.10.0,<2.0.0